backend: add update and draft feature for events
added an option to edit events
added an option to create event drafts
