+++ backend/src/middleware/validate.js
@@ -88,6 +88,30 @@ export const eventSchema = z.object({
   ).optional()
 });
 
+export const eventDraftSchema = z.object({
+  name: z.string().min(3).max(255),
+  type: z.enum(['HACKATHON', 'GENERAL_EVENT']),
+  tagline: z.string().optional().nullable(),
+  about: z.string().optional().nullable(),
+  maxParticipants: z.number().positive().optional().nullable(),
+  minTeamSize: z.number().positive().optional().nullable(),
+  maxTeamSize: z.number().positive().optional().nullable(),
+  eventTimeline: z.object({
+    timezone: z.string(),
+    applicationsStart: z.string().datetime(),
+    applicationsEnd: z.string().datetime(),
+    eventStart: z.string().datetime(),
+    eventEnd: z.string().datetime(),
+    rsvpDeadlineDays: z.number().nullable()
+  }),
+  eventLinks: z.object({
+    websiteUrl: z.string().url().optional().nullable(),
+    micrositeUrl: z.string().url().optional().nullable(),
+    contactEmail: z.string().email().optional().nullable(),
+    codeOfConductUrl: z.string().url().optional().nullable()
+  }).optional()
+});
+
 export const teamSchema = z.object({
   eventId: z.number(),
   name: z.string().min(3).max(255)
