+++ frontend/src/App.jsx
@@ -10,6 +10,7 @@ import Events from './pages/Events';
 import CreateEvent from './components/CreateEvent';
 import Profile from './pages/Profile';
 import EventDetails from './pages/EventDetails';
+import EditEvent from './components/EditEvent';
 
 const ProtectedRoute = ({ children, requiredRole }) => {
   const { user, loading } = useAuth();
@@ -49,6 +50,7 @@ const App = () => {
               <Route path="/register" element={<Register />} />
               <Route path="/events" element={<ProtectedRoute><Events /></ProtectedRoute>} />
               <Route path="/events/:id" element={<ProtectedRoute><EventDetails /></ProtectedRoute>} />
+              <Route path="/edit-event/:id" element={<ProtectedRoute><EditEvent /></ProtectedRoute>} />
               <Route path="/create-event" element={
                 <ProtectedRoute requiredRole="ADMIN">
                   <CreateEvent />
