+++ frontend/src/components/EditEvent.jsx
@@ -1,176 +1,78 @@
 import React, { useState, useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import { useAuth } from '../contexts/AuthContext';
-import api from '../utils/api';
+import api from '../api/api';
+import axios from 'axios';
+import EventForm from './EventForm';
 
 const EditEvent = () => {
   const { id } = useParams();
   const navigate = useNavigate();
   const { user } = useAuth();
+  const [localFormData, setEvent] = useState(null);
+  const [loading, setLoading] = useState(true);
   const [error, setError] = useState('');
-  const [formData, setFormData] = useState({
-    name: '',
-    type: 'hackathon',
-    tagline: '',
-    about: '',
-    max_participants: '',
-    min_team_size: '',
-    max_team_size: '',
-  });
 
   useEffect(() => {
     const fetchEvent = async () => {
       try {
-        const response = await api.get(`/api/events/${id}`);
-        setFormData(response.data);
+        const response = await axios.get(`${import.meta.env.VITE_API_URL}/api/events/${id}`, { withCredentials: true });
+        setEvent(response.data);
       } catch (error) {
-        setError('Failed to fetch event details');
-        console.error('Error:', error);
+        console.error('Error fetching event:', error);
+        setError('Failed to load event details');
+      } finally {
+        setLoading(false);
       }
     };
+
     fetchEvent();
   }, [id]);
 
-  const handleChange = (e) => {
-    setFormData({ ...formData, [e.target.name]: e.target.value });
-  };
+  if (loading) return <div>Loading...</div>;
+  if (error) return <div>{error}</div>;
+  if (!localFormData) return <div>Event not found</div>;
 
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    try {
-      await api.put(`/api/events/${id}`, formData);
-      navigate(`/events/${id}`);
-    } catch (error) {
-      setError('Failed to update event');
-      console.error('Error:', error);
-    }
-  };
-
-  if (user?.role !== 'organizer') {
+  if (user?.role !== 'ORGANIZER' || user?.id !== localFormData.createdById) {
     return <div className="text-center mt-8">You don't have permission to edit events.</div>;
   }
 
+  const apiCall = async (payload) => {
+    try {
+      const response = await axios.put(`${import.meta.env.VITE_API_URL}/api/events/${id}`, payload, { withCredentials: true });
+      console.log('Event updated:', response.data);
+      navigate('/events');
+    } catch (error) {
+      console.error('Error updating event:', error);
+      setError(error.response?.data?.error || 'An error occurred while updating the event. Please try again.');
+    }
+  }
+
+  localFormData.eventTimeline.eventStart = (localFormData.eventTimeline.eventStart).split('.')[0];
+  localFormData.eventTimeline.eventEnd = (localFormData.eventTimeline.eventEnd).split('.')[0];
+  localFormData.eventTimeline.applicationsStart = (localFormData.eventTimeline.applicationsStart).split('.')[0];
+  localFormData.eventTimeline.applicationsEnd = (localFormData.eventTimeline.applicationsEnd).split('.')[0];
+
+  if (!localFormData.eventLinks) {
+    localFormData.eventLinks = [
+      {
+        websiteUrl: '',
+        micrositeUrl: '',
+        contactEmail: '',
+        codeOfConductUrl: '',
+        socialLinks: null,
+      },
+    ];
+  }
+
   return (
-    <div className="max-w-7xl mx-auto py-6 sm:px-6 lg:px-8">
-      <div className="px-4 py-6 sm:px-0">
-        <h1 className="text-2xl font-semibold text-gray-900">Edit Event</h1>
-        {error && <p className="mt-2 text-sm text-red-600">{error}</p>}
-        <form onSubmit={handleSubmit} className="mt-5 space-y-6">
-          <div>
-            <label htmlFor="name" className="block text-sm font-medium text-gray-700">
-              Event Name
-            </label>
-            <input
-              type="text"
-              name="name"
-              id="name"
-              required
-              value={formData.name}
-              onChange={handleChange}
-              className="mt-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-            />
-          </div>
-
-          <div>
-            <label htmlFor="type" className="block text-sm font-medium text-gray-700">
-              Event Type
-            </label>
-            <select
-              id="type"
-              name="type"
-              value={formData.type}
-              onChange={handleChange}
-              className="mt-1 block w-full pl-3 pr-10 py-2 text-base border-gray-300 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm rounded-md"
-            >
-              <option value="hackathon">Hackathon</option>
-              <option value="general_event">General Event</option>
-            </select>
-          </div>
-
-          <div>
-            <label htmlFor="tagline" className="block text-sm font-medium text-gray-700">
-              Tagline
-            </label>
-            <input
-              type="text"
-              name="tagline"
-              id="tagline"
-              value={formData.tagline}
-              onChange={handleChange}
-              className="mt-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-            />
-          </div>
-
-          <div>
-            <label htmlFor="about" className="block text-sm font-medium text-gray-700">
-              About
-            </label>
-            <textarea
-              id="about"
-              name="about"
-              rows={3}
-              value={formData.about}
-              onChange={handleChange}
-              className="mt-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-            />
-          </div>
-
-          <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
-            <div>
-              <label htmlFor="max_participants" className="block text-sm font-medium text-gray-700">
-                Max Participants
-              </label>
-              <input
-                type="number"
-                name="max_participants"
-                id="max_participants"
-                value={formData.max_participants}
-                onChange={handleChange}
-                className="mt-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-              />
-            </div>
-
-            <div>
-              <label htmlFor="min_team_size" className="block text-sm font-medium text-gray-700">
-                Min Team Size
-              </label>
-              <input
-                type="number"
-                name="min_team_size"
-                id="min_team_size"
-                value={formData.min_team_size}
-                onChange={handleChange}
-                className="mt-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-              />
-            </div>
-
-            <div>
-              <label htmlFor="max_team_size" className="block text-sm font-medium text-gray-700">
-                Max Team Size
-              </label>
-              <input
-                type="number"
-                name="max_team_size"
-                id="max_team_size"
-                value={formData.max_team_size}
-                onChange={handleChange}
-                className="mt-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-              />
-            </div>
-          </div>
-
-          <div>
-            <button
-              type="submit"
-              className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-            >
-              Update Event
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
+    <EventForm
+    mode = {2} //to specify event edit mode
+    localFormData = {localFormData}
+    apiCall = {apiCall}
+    />
   );
+    
 };
 
 export default EditEvent;
